Upgrade github.com/cenk/backoff to v4
Import path has changed to github.com/cenkalti/backoff/v4.
The codebase has the assumption that the max backoff time is always
reached when set and the action has not succeeded.
This is no longer the case. When the backoff interval is going to exceed
the max backoff time it gives up immediately.
A quick fix (sleep) has been applied but the assumption should be
removed.