Add quit request timeout
To ensure scuttle does not hang on sending quit requests (especially
relevant if also providing a generic quit endpoint) a timeout can be
provided. When a request fails or times out, it is logged (if enabled)
but scuttle will exit cleanly and timely.
By default, the QUIT_REQUEST_TIMEOUT is set to 5 seconds.

When multiple generic quit endpoints are provided they are now called in
parallel to ensure the timeout is guaranteeing a quick shutdown of
scuttle.