Bugfix: Artifacts and Docker images are not build with same Go version
The release artifacts are build against whatever Go version
`ubuntu-latest` GitHub runner has. This happens to be 1.20 at the
moment, hopefully with the latest security fixes.
By using the newly introduced `GO_VERSION` env var in both
actions/setup-go and the docker build-args the versions are guaranteed
to be the same.